# cut: optimizations