… Cargo.lock
… src/uu/cut/src/buffer.rs