@sylvestre 2021-04-25
Well done for the huge wins!

Could you please create a benchmark files like:
https://github.com/uutils/coreutils/blob/master/src/uu/sort/BENCHMARKING.md
to be able to store way to benchmark this program? thanks