@cbjadwani 2021-04-28
Added benchmarking doc. I think it would be more accessible if instead of adding all details for each utility separately common details can be extracted and kept at project root level.